requirements.txt
requests == 2.24.0
bs4 == 0.0.1
pymongo == 3.10.1
kafka-python == 2.0.1
schedule == 0.6.0
